requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.25.2
requests==2.31.0

# Databases
motor==3.3.2
pymongo==4.6.0
elasticsearch==8.11.0

# Advanced NLP and ML
nltk==3.8.1
sumy==0.11.0
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
python-magic==0.4.27

# TTS and Audio
gTTS==2.4.0
pydub==0.25.1

# File handling and validation
aiofiles==23.2.1
python-magic-bin==0.4.14

# Additional dependencies
email-validator==2.1.0
python-multipart==0.0.6
jinja2==3.1.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1